// src/api/generated/schemas/ticketResponseStatus.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente
 * OpenAPI spec version: v1.0.0
 */

export type TicketResponseStatus = typeof TicketResponseStatus[keyof typeof TicketResponseStatus];


export const TicketResponseStatus = {
  PENDING: 'PENDING',
  AWAITING_CUSTOMER: 'AWAITING_CUSTOMER',
  AWAITING_DEVELOPMENT: 'AWAITING_DEVELOPMENT',
  COMPLETED: 'COMPLETED',
} as const;